requirements.txt
chainlit
psycopg2-binary
python-dotenv
openai